refactor(ai): isolate abortable retry sleep
